service/requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
pypdf2
numpy
python-multipart
